Correction des parametres en ligne de commande

+++ src/Asserv.Esial/config/params.h
@@ -124,7 +124,7 @@ PARAM(float, frontParMetreCodeurD)
 
 // Distance entre les roues codeuses en mm
 // !!! ATTENTION : redémarrer l'asserv pour prendre en compte les changements !!
-PARAM(int32_t, distRoues)
+PARAM(float, distRoues)
 
 // Nombre d'UO pour un tic de codeur
 // !!! ATTENTION : redémarrer l'asserv pour prendre en compte les changements !!
